server/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
tensorflow==2.18.0
python-multipart
pillow==10.1.0
numpy==1.26.0
opencv-python-headless>=4.5.0
pydantic==2.5.2
lime
scikit-image
aioredis>=2.0.0
fastapi-cache2>=0.2.0
redis==4.5.5
pydicom==2.4.3